common: Change authList v field to yParity in Transaction type (#806)
* common: Change authList v field to yParity in Transaction type

EIP-7702 defines it as y_parity. This has no actual effect on rlp
encoding/decoding. But the type is also used for JSON encoding
in execution API. There it is specifically defined as yParity and
thus for JSON-RPC requests it would not encode/decode it properly
(= always 0 v value without noticable error message)

* Make yParity uint8