discv5: remove unused listenAddress template (#823)
This is not used in any of our projects and more importantly I
believe this auto-selection does not need to be part of discv5.

Either is gets done beforehand and passed along discv5 init or
it gets handled by chronos, which is exactly how it is used
currently in the projects.